ccwrap.c: Added extra -isystem include path for C++
GCC 5.3 installs 2 sets of headers, the additional one only contains
a bits/ subdirectory with some important details like c++config.h.
(the subdirectory is actually named with the host triplet by we
create a symlink to that to maintain simplicity in the ccwrap program).

This should be harmless for builds of other GCCs which do not
include the extra subdirectory.