musl.build: Double install for GCC 5.3
This is an ugly hack but allows us to avoid the whole --with-sysroot
approach to building binutils & gcc.

We want musl installed in the STAGE_DIR directly for later use with
the ccwrap, and gcc wont find it while building unless it's installed
under STAGE_DIR/CROSS_TARGET.

Probably we can get away with removing STAGE_DIR/CROSS_TARGET entirely
after everything is built.