functions.sh: Special handling of CROSS_TARGET
When building GCC 5.3 and musl, it is important to
set the triplet to something that is -linux-musl*

The approach we take here is to assume first that
the CROSS_TARGET is setup with something with -gnu
in the last portion, and substitute that for -musl
in the case we are building with GCC 5.3 and are
using the musl libc implementation.

This allows the targets to specify a CROSS_TARGET
with the 'gnu' word, such as 'armv5l-unknown-linux-gnueabi'
and that will automatically translate to the
correct triplet 'armv5l-unknown-linux-gnueabi'.

Targets which fail to specify the trailing '-gnu' will be
problematic but should build fine with the added '-gnu'.